+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.7] - 2020-12-01
 ### Added
 - Logger for store http requests into logs/http.log.
 - Middleware for log http requests.
@@ -12,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Npm script for stop process using the pid stored in .pid file.
 - SGTERM and SIGINT signal handlers.
 - Implement API call for FWCloud Websrv updates (PUT /updates/websrv).
+- HTTPS support.
 
 ## Changed
 - Improved compare versions error handling.